+++ gossip/old_scrambler/tx_scrambler.go
